backend/requirements.txt
# Django Core
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-filter==23.5

# File handling & SMB
smbprotocol==1.12.0
# python-magic==0.4.27  # Comentado - no necesario para producción
# python-magic-bin solo es para Windows, no se usa en Docker Linux

# Background tasks
celery==5.3.6
django-celery-beat==2.5.0

# AI & Machine Learning
groq==0.34.1
httpx==0.27.2  # Requerido por groq

# Utilities
python-dateutil==2.8.2
Pillow==10.2.0
pytz==2024.1
requests==2.31.0
python-dotenv==1.0.0

# Microsoft Integration
msal==1.24.0

# OnlyOffice Integration
PyJWT==2.8.0

# Development
ipython==8.20.0
django-extensions==3.2.3

# Production
gunicorn==21.2.0
whitenoise==6.6.0


# Excel file handling
openpyxl